+++ sharding-jdbc/src/main/java/io/shardingsphere/core/api/MasterSlaveDataSourceFactory.java
@@ -19,13 +19,10 @@ package io.shardingsphere.core.api;
 
 import io.shardingsphere.core.api.config.MasterSlaveRuleConfiguration;
 import io.shardingsphere.core.jdbc.core.datasource.MasterSlaveDataSource;
-import io.shardingsphere.core.yaml.masterslave.YamlMasterSlaveConfiguration;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 
 import javax.sql.DataSource;
-import java.io.File;
-import java.io.IOException;
 import java.sql.SQLException;
 import java.util.Map;
 
@@ -40,8 +37,6 @@ public final class MasterSlaveDataSourceFactory {
     /**
      * Create master-slave data source.
      *
-     * <p>One master data source can configure multiple slave data source.</p>
-     *
      * @param dataSourceMap data source map
      * @param masterSlaveRuleConfig master-slave rule configuration
      * @param configMap configuration map
@@ -51,67 +46,5 @@ public final class MasterSlaveDataSourceFactory {
     public static DataSource createDataSource(
             final Map<String, DataSource> dataSourceMap, final MasterSlaveRuleConfiguration masterSlaveRuleConfig, final Map<String, Object> configMap) throws SQLException {
         return new MasterSlaveDataSource(dataSourceMap, masterSlaveRuleConfig, configMap);
-    }
-    
-    /**
-     * Create master-slave data source.
-     *
-     * <p>One master data source can configure multiple slave data source.</p>
-     *
-     * @param yamlFile yaml file for master-slave rule configuration with data sources
-     * @return master-slave data source
-     * @throws SQLException SQL exception
-     * @throws IOException IO exception
-     */
-    public static DataSource createDataSource(final File yamlFile) throws SQLException, IOException {
-        YamlMasterSlaveConfiguration config = YamlMasterSlaveConfiguration.unmarshal(yamlFile);
-        return createDataSource(config.getDataSources(), config.getMasterSlaveRule().getMasterSlaveRuleConfiguration(), config.getMasterSlaveRule().getConfigMap());
-    }
-    
-    /**
-     * Create master-slave data source.
-     *
-     * <p>One master data source can configure multiple slave data source.</p>
-     *
-     * @param yamlBytes yaml bytes for master-slave rule configuration with data sources
-     * @return master-slave data source
-     * @throws SQLException SQL exception
-     * @throws IOException IO exception
-     */
-    public static DataSource createDataSource(final byte[] yamlBytes) throws SQLException, IOException {
-        YamlMasterSlaveConfiguration config = YamlMasterSlaveConfiguration.unmarshal(yamlBytes);
-        return createDataSource(config.getDataSources(), config.getMasterSlaveRule().getMasterSlaveRuleConfiguration(), config.getMasterSlaveRule().getConfigMap());
-    }
-    
-    /**
-     * Create master-slave data source.
-     *
-     * <p>One master data source can configure multiple slave data source.</p>
-     *
-     * @param dataSourceMap data source map
-     * @param yamlFile yaml file for master-slave rule configuration without data sources
-     * @return master-slave data source
-     * @throws SQLException SQL exception
-     * @throws IOException IO exception
-     */
-    public static DataSource createDataSource(final Map<String, DataSource> dataSourceMap, final File yamlFile) throws SQLException, IOException {
-        YamlMasterSlaveConfiguration config = YamlMasterSlaveConfiguration.unmarshal(yamlFile);
-        return createDataSource(dataSourceMap, config.getMasterSlaveRule().getMasterSlaveRuleConfiguration(), config.getMasterSlaveRule().getConfigMap());
-    }
-    
-    /**
-     * Create master-slave data source.
-     *
-     * <p>One master data source can configure multiple slave data source.</p>
-     *
-     * @param dataSourceMap data source map
-     * @param yamlBytes yaml bytes for master-slave rule configuration without data sources
-     * @return master-slave data source
-     * @throws SQLException SQL exception
-     * @throws IOException IO exception
-     */
-    public static DataSource createDataSource(final Map<String, DataSource> dataSourceMap, final byte[] yamlBytes) throws SQLException, IOException {
-        YamlMasterSlaveConfiguration config = YamlMasterSlaveConfiguration.unmarshal(yamlBytes);
-        return createDataSource(dataSourceMap, config.getMasterSlaveRule().getMasterSlaveRuleConfiguration(), config.getMasterSlaveRule().getConfigMap());
     }
 }
